+++ utils/functions/initData.ts
