+++ utils/functions/updateNeedToStudy.ts
